+++ lib/ember_cli/path_set.rb
@@ -19,10 +19,6 @@ module EmberCli
       end
     end
 
-    def index_file
-      dist.join("index.html")
-    end
-
     def tmp
       @tmp ||= root.join("tmp").tap(&:mkpath)
     end