+++ spec/lib/ember_cli/path_set_spec.rb
@@ -52,17 +52,6 @@ describe EmberCli::PathSet do
     end
   end
 
-  describe "#index_file" do
-    it "depends on the app name" do
-      app = build_app(name: "foo")
-
-      path_set = build_path_set(app: app)
-
-      expect(path_set.index_file).
-        to eq ember_cli_root.join("apps", "foo", "index.html")
-    end
-  end
-
   describe "#gemfile" do
     it "is a child of #root" do
       app = build_app(name: "foo")