+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 master
 ------
 
+0.7.0
+-----
+
 * Introduce the `deploy` configuration option.
 * Translate Rails environments other than `test` or `development` to
   `production`, unless an `EMBER_ENV` is specified. [#366]